+++ backend/app/models/collection.py
@@ -1,15 +1,17 @@
+import enum
 from uuid import UUID, uuid4
 from datetime import datetime
-from typing import Optional
+from typing import Any, List, Optional
+from dataclasses import dataclass, field, fields
 
 from sqlmodel import Field, Relationship, SQLModel
+from pydantic import HttpUrl, BaseModel
 
 from app.core.util import now
-from .user import User
 from .organization import Organization
 from .project import Project
-import enum
-from enum import Enum
+from app.core.util import now
+from app.crud.document import DocumentCrud
 
 
 class CollectionStatus(str, enum.Enum):
@@ -20,12 +22,6 @@ class CollectionStatus(str, enum.Enum):
 
 class Collection(SQLModel, table=True):
     id: UUID = Field(default_factory=uuid4, primary_key=True)
-
-    owner_id: int = Field(
-        foreign_key="user.id",
-        nullable=False,
-        ondelete="CASCADE",
-    )
 
     organization_id: int = Field(
         foreign_key="organization.id",
@@ -44,11 +40,104 @@ class Collection(SQLModel, table=True):
 
     status: CollectionStatus = Field(default=CollectionStatus.processing)
     error_message: Optional[str] = Field(default=None, nullable=True)
+    task_id: Optional[str] = Field(default=None, description="Celery task ID")
 
     created_at: datetime = Field(default_factory=now)
     updated_at: datetime = Field(default_factory=now)
     deleted_at: Optional[datetime] = None
 
-    owner: User = Relationship(back_populates="collections")
     organization: Organization = Relationship(back_populates="collections")
     project: Project = Relationship(back_populates="collections")
+
+
+@dataclass
+class ResponsePayload:
+    status: str
+    route: str
+    key: str = field(default_factory=lambda: str(uuid4()))
+    time: str = field(default_factory=lambda: now().strftime("%c"))
+
+    @classmethod
+    def now(cls):
+        attr = "time"
+        for i in fields(cls):
+            if i.name == attr:
+                return i.default_factory()
+
+        raise AttributeError(f'Expected attribute "{attr}" does not exist')
+
+
+# pydantic models
+class DocumentOptions(BaseModel):
+    documents: List[UUID] = Field(
+        description="List of document IDs",
+    )
+    batch_size: int = Field(
+        default=1,
+        description=(
+            "Number of documents to send to OpenAI in a single "
+            "transaction. See the `file_ids` parameter in the "
+            "vector store [create batch](https://platform.openai.com/docs/api-reference/vector-stores-file-batches/createBatch)."
+        ),
+    )
+
+    def model_post_init(self, __context: Any):
+        self.documents = list(set(self.documents))
+
+    def __call__(self, crud: DocumentCrud):
+        (start, stop) = (0, self.batch_size)
+        while True:
+            view = self.documents[start:stop]
+            if not view:
+                break
+            yield crud.read_each(view)
+            start = stop
+            stop += self.batch_size
+
+
+class AssistantOptions(BaseModel):
+    # Fields to be passed along to OpenAI. They must be a subset of
+    # parameters accepted by the OpenAI.clien.beta.assistants.create
+    # API.
+    model: str = Field(
+        description=(
+            "OpenAI model to attach to this assistant. The model "
+            "must compatable with the assistants API; see the "
+            "OpenAI [model documentation](https://platform.openai.com/docs/models/compare) for more."
+        ),
+    )
+    instructions: str = Field(
+        description=(
+            "Assistant instruction. Sometimes referred to as the " '"system" prompt.'
+        ),
+    )
+    temperature: float = Field(
+        default=1e-6,
+        description=(
+            "Model temperature. The default is slightly "
+            "greater-than zero because it is [unknown how OpenAI "
+            "handles zero](https://community.openai.com/t/clarifications-on-setting-temperature-0/886447/5)."
+        ),
+    )
+
+
+class CallbackRequest(BaseModel):
+    callback_url: Optional[HttpUrl] = Field(
+        default=None,
+        description="URL to call to report endpoint status",
+    )
+
+
+class CreationRequest(
+    DocumentOptions,
+    AssistantOptions,
+    CallbackRequest,
+):
+    def extract_super_type(self, cls: "CreationRequest"):
+        for field_name in cls.__fields__.keys():
+            field_value = getattr(self, field_name)
+            yield (field_name, field_value)
+
+
+class DeletionRequest(CallbackRequest):
+    collection_id: UUID = Field("Collection to delete")
